+++ package.json
@@ -7,7 +7,7 @@
   },
   "type": "module",
   "scripts": {
-    "start": "node src/index.js"
+    "start": "node src/task2.js"
   },
   "repository": {
     "type": "git",
