+++ src/task2-modules/handleOSCommand.js
@@ -1,0 +1,28 @@
+import os from 'os';
+import process from 'process';
+
+export async function handleOSCommand(flag) {
+    switch (flag) {
+        case '--EOL':
+            console.log(JSON.stringify(os.EOL));
+            break;
+        case '--cpus':
+            const cpus = os.cpus();
+            console.log(`Total CPUs: ${cpus.length}`);
+            cpus.forEach((cpu, index) => {
+                console.log(`CPU ${index + 1}: ${cpu.model}, ${cpu.speed / 1000} GHz`);
+            });
+            break;
+        case '--homedir':
+            console.log(os.homedir());
+            break;
+        case '--username':
+            console.log(os.userInfo().username);
+            break;
+        case '--architecture':
+            console.log(process.arch);
+            break;
+        default:
+            console.log('Invalid input');
+    }
+}
