+++ src/task2-modules/printCurrentDirectory.js
@@ -1,0 +1,3 @@
+export function printCurrentDirectory(currentDir) {
+    console.log('\x1b[32m%s\x1b[0m', `You are currently in ${currentDir}`);
+}
